feat: add function display_syntax
