+++ src/envuse/create_program.rs
@@ -1,9 +1,6 @@
-use crate::errors::program_error::ProgramError;
-
-use crate::syntax_error::SyntaxError;
-
 use super::super::parser::ast::AST;
 use super::super::parser::tokenizer::Tokenizer;
+use super::display_program_error::display_program_error;
 use super::program::Program;
 
 pub fn create_program<T: ToString>(
@@ -20,17 +17,5 @@ pub fn create_program<T: ToString>(
         }
     };
 
-    match result_program {
-        Err(error) if error.is::<SyntaxError>() => {
-            let syntax_error = error.downcast::<SyntaxError>().unwrap();
-            do yeet ProgramError {
-                message: format!("SyntaxError: {}", syntax_error.message),
-                span: Some(syntax_error.span.clone()),
-                source: source.to_string(),
-                location: location_val,
-                cause: Some(*syntax_error),
-            }
-        }
-        result_program => result_program,
-    }
+    display_program_error(result_program, source, location_val)
 }
