+++ src/lib.rs
@@ -9,4 +9,5 @@ pub mod errors;
 pub mod parser;
 pub mod syntax_error;
 pub mod transformers;
+pub mod utils;
 pub use envuse::create_program::create_program;
